pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vs-waypoint-macros"
version = "1.0.0"
description = "Generate waypoint macros for Vintage Story with visual numpad reference"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Keenan" }
]
keywords = ["vintage-story", "waypoints", "macros", "gaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
]
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[project.scripts]
vs-waypoint-macros = "vs_waypoint_macros.cli:main"

[project.urls]
Homepage = "https://github.com/keenan/vintagestory-waypoint-macros"
Repository = "https://github.com/keenan/vintagestory-waypoint-macros"
Issues = "https://github.com/keenan/vintagestory-waypoint-macros/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/vs_waypoint_macros"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["vs_waypoint_macros"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
